fix guided mode position controller in XY and Z mode for our use-case
In guided mode, the position controller now assumes that desired velocities are
received with high frequency from the controller and that these should be
applied as-is without any intermediate corrections to the desired
position. This applies both to the XY and the Z controller. Without this
commit, PixHawk's own internal corrections would interfere with the
desired velocities we send from the companion computer.