ensure that the RTCM protocol is enabled for the u-Blox GPS after startup
I am not sure whether this is actually needed for our drones; in theory,
the uBlox GPS devices should already be configured to accept RTCM
messages. Nevertheless, it doesn't hurt to have this patch in the code.

In the future, we should test whether this patch is really necessary and
get rid of it if it isn't needed.